requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-groq==0.1.3
langchain-objectbox
pypdf==4.2.0
pdfplumber==0.11.5
python-dotenv==1.0.1
pytesseract==0.3.13
sentence_transformers==3.3.1
fastapi==0.115.6
uvicorn[standard]==0.32.0
python-multipart==0.0.9
Pillow==11.0.0